Flask Deployed App/requirements.txt
absl-py==2.3.1
altair==5.5.0
catboost==1.2.8
category_encoders==2.8.1
dill==0.4.0
filetype==1.2.0
Flask==3.0.3
flask-cors==6.0.1
fsspec==2025.10.0
gitpython==3.1.45
graphviz==0.21
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
ImageHash==4.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.2
kaggle==1.7.4.5
keras==3.12.0
libclang==18.1.1
Markdown==3.10
matplotlib==3.10.0
missingno==0.5.2
ml-dtypes==0.5.3
networkx==3.5
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
opt-einsum==3.4.0
optree==0.17.0
pandas==2.2.3
pandas-profiling==3.2.0
phik==0.12.5
pillow==10.4.0
plotly==6.3.1
protobuf==6.33.0
psutil==6.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic-core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
regex==2025.9.1
requests==2.32.3
rich==14.2.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
shap==0.48.0
statsmodels==0.14.4
streamlit==1.51.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.66.5
typing-extensions==4.12.2
Werkzeug==3.0.4
wheel==0.45.1
wordcloud==1.9.4
xgboost==3.0.5
ydata-profiling==4.17.0
